feat: update next.config to new version of nextra
